Wrapper default ConfigurationResolver to reduce complexity
