+++ subprojects/dependency-management/src/main/java/org/gradle/api/internal/artifacts/DefaultDependencyManagementServices.java
@@ -39,6 +39,7 @@ import org.gradle.api.internal.artifacts.dsl.dependencies.DependencyFactory;
 import org.gradle.api.internal.artifacts.dsl.dependencies.ProjectFinder;
 import org.gradle.api.internal.artifacts.ivyservice.CacheLockingManager;
 import org.gradle.api.internal.artifacts.ivyservice.DefaultConfigurationResolver;
+import org.gradle.api.internal.artifacts.ivyservice.DependencyLockingConfigurationResolver;
 import org.gradle.api.internal.artifacts.ivyservice.ErrorHandlingConfigurationResolver;
 import org.gradle.api.internal.artifacts.ivyservice.IvyContextManager;
 import org.gradle.api.internal.artifacts.ivyservice.IvyContextualArtifactPublisher;
@@ -242,22 +243,23 @@ public class DefaultDependencyManagementServices implements DependencyManagement
                                                        DependencyLockManager dependencyLockManager) {
             return new ErrorHandlingConfigurationResolver(
                     new ShortCircuitEmptyConfigurationResolver(
-                        new DefaultConfigurationResolver(
-                            artifactDependencyResolver,
-                            repositories,
-                            metadataHandler,
-                            resolutionResultsStoreFactory,
-                            startParameter.isBuildProjectDependencies(),
-                            attributesSchema,
-                            new DefaultArtifactTransforms(
-                                new VariantAttributeMatchingCache(
-                                    variantTransforms,
-                                    attributesSchema,
-                                    attributesFactory),
-                                attributesSchema),
-                            moduleIdentifierFactory,
-                            buildOperationExecutor,
-                            artifactTypeRegistry,
+                        new DependencyLockingConfigurationResolver(
+                            new DefaultConfigurationResolver(
+                                artifactDependencyResolver,
+                                repositories,
+                                metadataHandler,
+                                resolutionResultsStoreFactory,
+                                startParameter.isBuildProjectDependencies(),
+                                attributesSchema,
+                                new DefaultArtifactTransforms(
+                                    new VariantAttributeMatchingCache(
+                                        variantTransforms,
+                                        attributesSchema,
+                                        attributesFactory),
+                                    attributesSchema),
+                                moduleIdentifierFactory,
+                                buildOperationExecutor,
+                                artifactTypeRegistry),
                             projectFinder,
                             startParameter.isDependencyLockEnabled(),
                             dependencyLockManager),
